server/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "AI Travel Copilot using MCP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.51.0",
    "mcp[cli,tcp,ws]==1.8.0",
    "nest-asyncio>=1.6.0",
    "prompt-toolkit>=3.0.51",
    "python-dotenv>=1.1.0",
    "streamlit>=1.55.0",
]

